requirements.txt
requests
streamlit
psycopg2-binary
python-dotenv
supabase
google-generativeai
ipython